+++ docs/pages/02-basic-usage.md
@@ -11,10 +11,10 @@ the class definition. The four main decorators are:
 - `@Expose` specifies how expose the given property on the plain object
 - `@Exclude` marks the property as skipped, so it won't show up in the transformation
 - `@Transform` allows specifying a custom transformation on the property via a custom handler
-- `@Nested` decorator explicitly sets the type of the property, during the transformation `class-transformer` will attempt
+- `@Nested` decorator explicitly sets the type of the property, during the transformation `class-transform` will attempt
   to create an instance of the specified type
 
 You must always decorate all your properties with an `@Expose` or `@Exclude` decorator.
 
-> **NOTE:** It's important to remember `class-transformer` will call the target type with am empty constructor, so if
+> **NOTE:** It's important to remember `class-transform` will call the target type with am empty constructor, so if
 > you are using a type what requires special setup, you need to use a `@Transform` decorator and create the instance yourself.
