+++ src/constants/default-options.constant.ts
@@ -1,4 +1,4 @@
-import { ClassTransformOptions } from '../interfaces/class-transformer-options.interface';
+import { ClassTransformOptions } from '../interfaces/class-transform-options.interface';
 
 /**
  * These are the default options used by any transformation operation.
