+++ src/interfaces/index.ts
@@ -9,5 +9,5 @@ export * from './metadata/transform-metadata.interface';
 export * from './metadata/transform-fn-params.interface';
 export * from './metadata/type-metadata.interface';
 export * from './class-constructor.type';
-export * from './class-transformer-options.interface';
+export * from './class-transform-options.interface';
 export * from './target-map.interface';
