+++ src/interfaces/metadata/transform-fn-params.interface.ts
@@ -1,5 +1,5 @@
 import { TransformationType } from '../../enums';
-import { ClassTransformOptions } from '../class-transformer-options.interface';
+import { ClassTransformOptions } from '../class-transform-options.interface';
 
 export interface TransformFnParams {
   value: any;
