+++ test/functional/transformer-order.spec.ts
@@ -18,7 +18,7 @@ describe('applying several transformations', () => {
     const firstUser = plainToInstance(User, { name: 'Joe' });
     expect(firstUser.name).toEqual('John');
 
-    // Prior to this pull request [#355](https://github.com/typestack/class-transformer/pull/355)
+    // Prior to this pull request [#355](https://github.com/cunarist/class-transform/pull/355)
     // the order of the transformations was reversed after every `plainToInstance()` call
     // So after consecutive calls `User#name` would be "John" - "Jonathan" - "John" - "Jonathan"...
     // This test ensures the last transformation is always the last one to be applied
